HDX-5817 - [tools.humdata.org]: wizard vers 2.0          - refactoring 3 step process into single step          - intentional left routing in place and first step so we can have compatibility with old url          - some adjustments to flow + new embed/share menu
