+++ src/app/app-routing.module.ts
@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { ImportComponent } from './tools-wizard/import/import.component';
-import { SelectComponent } from './tools-wizard/select/select.component';
-import { ShareComponent } from './tools-wizard/share/share.component';
 
 const routes: Routes = [
   {
@@ -10,20 +8,6 @@ const routes: Routes = [
     component: ImportComponent,
     data: {
       title: 'Import'
-    }
-  },
-  {
-    path: 'select',
-    component: SelectComponent,
-    data: {
-      title: 'Select'
-    }
-  },
-  {
-    path: 'share',
-    component: ShareComponent,
-    data: {
-      title: 'Share'
     }
   },
   { path: '**',
