+++ src/app/tools-wizard/import/import.component.less
@@ -4,6 +4,7 @@
 @sampleItemRadioSize: 20px;
 .sample-item {
   position: relative;
+  max-width: 800px;
   width: 100%;
   background-color: @whiteColor;
   margin-bottom: 10px;
@@ -50,3 +51,51 @@
     width: 100%;
   }
 }
+
+.quick-charts-render {
+  padding: 10px 20px;
+  background-color: @whiteColor;
+}
+.share-section {
+  ul {
+    list-style: none;
+    padding: 0;
+    a {
+      cursor: pointer;
+    }
+    a.disabled {
+      color: @lightGrayColor;
+      cursor: not-allowed;
+    }
+  }
+
+  @embedCodeFontSize: 16px;
+  @embedCodeLineHeight: 1.5*@embedCodeFontSize;
+  #embedCode {
+    margin-top: 10px;
+    .sourceSansProFont(400, @embedCodeFontSize);
+    width: 100%;
+    line-height: @embedCodeLineHeight;
+    height: 4*@embedCodeLineHeight + 2*@wizardFormTBPadding;
+  }
+}
+
+.vertical-menu {
+  list-style: none;
+  display: inline;
+
+  li {
+    display: inline;
+    &:after {
+      padding: 0 10px;
+      content: '|';
+    }
+    &:last-child {
+      margin-right: 0;
+      &:after {
+        content: none;
+      }
+    }
+  }
+}
+
