--- src/app/tools-wizard/select/select.component.html
@@ -1,76 +1,0 @@
-<div id="step2" class="step">
-  <div class="paddingRowHack step-header">
-    <div class="paddingLeftHack paddingRightHack">
-      <div class="row" style="position: relative;">
-        <div class="col-xs-8">
-          <div class="title">{{stepName}}</div>
-          <div class="description">Recipes determine what charts will be available in the next step. These are generated from the HXL hashtags in your data.</div>
-        </div>
-        <div class="col-xs-4">
-          <div class="btn-actions">
-            <button routerLink="/import" class="btn btn-empty-black">Previous</button>
-            <button (click)="navigateToShare()" [disabled]="totalBites() === 0" class="btn btn-primary next-step">Next Step</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-  <div class="row step-content">
-    <div class="col-xs-3">
-      <div class="radio-selection">
-        <div class="radio-title">Recipe options:</div>
-        <label for="recipe-opt-hdx">
-          <input type="radio" name="data-source" value="sample" id="recipe-opt-hdx" (change)="changeRecipe($event);" [checked]="getWizardConfig().step2Sample">HDX Quick Charts Recipes <br>(See <a href="https://github.com/OCHA-DAP/hxl-recipes" target="_blank" >a list of 100+ recipes</a>)
-        </label>
-        <label for="recipe-opt-custom">
-          <input type="radio" name="data-source" value="url" id="recipe-opt-custom" (change)="changeRecipe($event);" [checked]="!getWizardConfig().step2Sample"> Custom Recipes <br>
-          (Bring your own json file to create custom charts. Read <a target="_blank" href="https://github.com/OCHA-DAP/hxl-recipes">documentation</a> to learn more)
-          <ul>
-            <li>
-              <a href="javascript:void(0)" hdxDropboxchooser (urlSelect)="updateSelectedRecipeUrl($event)">
-                <img src="./assets/icons/dropbox.png"> Dropbox
-              </a>
-            </li>
-            <li>
-              <a href="javascript:void(0)" hdxGooglepicker (urlSelect)="updateSelectedRecipeUrl($event)">
-                <img src="./assets/icons/drive.png">Google Drive
-              </a>
-            </li>
-          </ul>
-        </label>
-      </div>
-    </div>
-    <div class="col-xs-9 recipe-results">
-      <div class="row" *ngIf="!getWizardConfig().step2Sample">
-          <div class="source-custom-url">
-            <label for="source-custom-recipe-url-field">Enter URL</label>
-            <input [ngModel]="getWizardConfig().recipeUrl" (change)="recipeUrlChanged($event)" name="selectedRecipeUrl" class="wizard-input" id="source-custom-recipe-url-field" placeholder="eg. http://example.com/population-jan-2011.json">
-          </div>
-      </div>
-      <div class="row">
-        <div class="results-title"> {{totalBites()}} charts are available based on the recipes<span class="results-title" *ngIf="totalBites() === 0">. Please choose another recipe or dataset.</span></div>
-
-        <div class="result-group" *ngIf="chartBites.length > 0">
-          <div class="section-title">Bar chart</div>
-          <ul class="result-list">
-            <li *ngFor="let b of chartBites">{{b.title}}</li>
-          </ul>
-        </div>
-
-        <div class="result-group" *ngIf="timeseriesBites.length > 0">
-          <div class="section-title">Line chart (Timeseries)</div>
-          <ul class="result-list">
-            <li *ngFor="let b of timeseriesBites">{{b.title}}</li>
-          </ul>
-        </div>
-
-        <div class="result-group" *ngIf="keyfigureBites.length > 0">
-          <div class="section-title">Key figures</div>
-          <ul class="result-list">
-            <li *ngFor="let b of keyfigureBites">{{b.title}}</li>
-          </ul>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
