--- src/app/tools-wizard/select/select.component.less
@@ -1,26 +1,0 @@
-@import "../step-style";
-
-.recipe-results {
-  .results-title {
-    .sourceSansProFont(600, 18px);
-    color: @darkGrayColor;
-    margin-bottom: 15px;
-  }
-
-  .result-group {
-    .section-title {
-      margin-top: 20px;
-    }
-    ul {
-      padding: 0;
-      list-style: none;
-
-
-      li {
-        .sourceSansProFont(400, 16px);
-        color: @darkGrayColor;
-        margin: 5px 0;
-      }
-    }
-  }
-}
