--- src/app/tools-wizard/share/share.component.html
@@ -1,69 +1,0 @@
-<div id="step3" class="step">
-  <div class="paddingRowHack step-header">
-    <div class="paddingLeftHack paddingRightHack">
-      <div class="row" style="position: relative;">
-        <div class="col-xs-9">
-          <div class="title">{{stepName}}</div>
-          <div class="description">Choose 3 default charts to display from the dropdown. Share Quick Charts for publication.</div>
-        </div>
-        <div class="col-xs-3">
-          <div class="btn-actions">
-            <button routerLink="/select" class="btn btn-empty-black">Previous</button>
-            <div class="btn-group" dropdown>
-              <button (click)="prepareShare($event);" type="button" class="btn btn-primary next-step dropdown-toggle">
-                Share
-              </button>
-              <!--
-              <button dropdownToggle type="button" class="btn btn-primary next-step dropdown-toggle">
-                Share <span class="caret"></span>
-              </button>
-              <ul *dropdownMenu class="dropdown-menu dropdown-menu-right" role="menu">
-                <li role="menuitem"><a class="dropdown-item" href disabled="disabled">Download as image</a></li>
-                <li role="menuitem"><a class="dropdown-item" (click)="embedCode.scrollIntoView();">Embed</a></li>
-              </ul>
-              -->
-            </div>
-          </div>
-
-        </div>
-
-      </div>
-    </div>
-  </div>
-  <div class="row step-content">
-    <div class="col-xs-12">
-      <div class="quick-charts-render">
-        <iframe id="quick-charts-iframe" [src]="iFrameUrl" (load)="iFrameLoaded()"
-                style="border:none; width:100%; min-height:520px">
-        </iframe>
-      </div>
-    </div>
-    <div class="col-xs-12">
-      <div class="data-source">
-        Data source: <a target="_blank" [href]="dataSource">{{ dataSource }}</a>
-      </div>
-    </div>
-    <div class="col-xs-12 share-section">
-      <div class="row">
-        <div class="col-xs-3">
-          <div class="section-title">Download:</div>
-          <ul>
-            <!--<li><a>Download as SVG</a></li>-->
-            <li><a (click)="prepareSnapshot($event)">Download as image</a></li>
-          </ul>
-          <div class="section-title">Push to HDX:</div>
-          <ul>
-            <li><a class="disabled">Publish data and charts to HDX (coming soon)</a></li>
-          </ul>
-        </div>
-        <div class="col-xs-9">
-          <div><span class="section-title">Embed (Live charts):</span> if you embed - charts will be live and always reflect the latest data from the source.</div>
-          <div>
-            <textarea id="embedCode" #embedCode class="wizard-input" (click)="prepareShare($event)" readonly [innerText]="embedUrl"></textarea>
-          </div>
-        </div>
-      </div>
-    </div>
-
-  </div>
-</div>
