--- src/app/tools-wizard/share/share.component.less
@@ -1,33 +1,0 @@
-@import "../step-style";
-@import "../form-styles";
-
-.quick-charts-render {
-  padding: 10px 20px;
-  background-color: @whiteColor;
-}
-.share-section {
-  margin-top: 30px;
-
-  ul {
-    list-style: none;
-    padding: 0;
-    a {
-      cursor: pointer;
-    }
-    a.disabled {
-      color: @lightGrayColor;
-      cursor: not-allowed;
-    }
-  }
-
-  @embedCodeFontSize: 16px;
-  @embedCodeLineHeight: 1.5*@embedCodeFontSize;
-  #embedCode {
-    margin-top: 10px;
-    .sourceSansProFont(400, @embedCodeFontSize);
-    width: 100%;
-    line-height: @embedCodeLineHeight;
-    height: 4*@embedCodeLineHeight + 2*@wizardFormTBPadding;
-  }
-}
-
