+++ src/app/tools-wizard/tools-wizard.component.less
@@ -3,18 +3,30 @@
 @wizardH1Size: 30px;
 @wizardH1LineHeight: 1.1;
 @wizardTBPadding: 20px;
-@wizardHeaderBorderSize: 2px;
+@wizardHeaderBorderSize: 1px;
+@wizardContainerPadding: 30px;
 
 .wizard-container {
   height: 100vh;
-  padding: 0 @containerMarginSize;
+  padding: 0 2*@containerMarginSize;
   margin: 0 -15px;
 }
 
+/deep/ .paddingDoubleRowHack {
+  margin: 0 -2*@containerMarginSize;
+  .paddingDoubleLRHack {
+    padding: 0 2*@containerMarginSize;
+  }
+}
+
+.wizard-header-container {
+  background-color: @whiteColor;
+}
 .wizard-header {
   //margin: 0 -15px;
   //padding: 0 15px;
   background-color: @whiteColor;
+  margin: 0 @containerMarginSize;
 
   border-bottom: @wizardHeaderBorderSize solid @lightGrayColor;
   h1 {
@@ -114,7 +126,7 @@ ol.steps {
   background-color: rgba(0, 0, 0, 0.7);
   z-index: 10000;
   position: fixed;
-  top: @wizardHeaderBorderSize + @stepsContainerHeight;
+  top: 0;//@wizardHeaderBorderSize + @stepsContainerHeight;
   left: 0;
   width: 100%;
   height: 100%;
@@ -122,7 +134,7 @@ ol.steps {
   .loading-content {
     position: absolute;
     left: 50%;
-    top: 47%; // :)
+    top: 50%;
     transform: translate(-50%, -50%);
 
     color: @whiteColor;
